Actualización2

+++ app/src/main/java/com/appclimaparcial/myapplication/presentacion/Ciudades/CiudadesViewModel.kt
@@ -45,7 +45,7 @@ class CiudadesViewModel(
         }
     }
 
-    private fun buscarPorGeolocalizacion(lat: Double, lon: Double) {
+    private fun buscarPorGeolocalizacion(lat: Float, lon: Float) {
         uiState = CiudadesEstado.Cargando
         viewModelScope.launch {
             try {
@@ -60,8 +60,11 @@ class CiudadesViewModel(
     }
 
     private fun seleccionar(ciudad: Ciudad){
-        userPrefs.saveCiudadSeleccionada(ciudad) // Para que siga funcionando lo de geolocalizacion
-        val ruta = Ruta.Clima(lat = ciudad.lat, lon = ciudad.lon, nombre = ciudad.name
+        userPrefs.saveCiudadSeleccionada(ciudad)
+        val ruta = Ruta.Clima(
+            lat = ciudad.lat,
+            lon = ciudad.lon,
+            nombre = ciudad.name
         )
         router.navegar(ruta)
     }
