+++ app/src/main/java/com/appclimaparcial/myapplication/repository/Repositorio.kt
@@ -6,7 +6,7 @@ import com.appclimaparcial.myapplication.repository.modelos.ListForecast
 
 interface Repositorio {
     suspend fun buscarCiudad(ciudad: String): List<Ciudad>
-    suspend fun buscarCiudadPorCoords(lat: Double, lon: Double): List<Ciudad>  // Agregamos acá la búqueda por coordenadas
+    suspend fun buscarCiudadPorCoords(lat: Float, lon: Float): List<Ciudad>  // Agregamos acá la búqueda por coordenadas
     suspend fun traerClima(lat: Float, lon: Float) : Clima
     suspend fun traerPronostico(nombre: String) : List<ListForecast>
 }
